Complete linux modify routine
Modify works in linux, works in router. Handles directories as
intended, and handles cleanups correctly as tested. Creates a
unique filename at each pass as originally intended. We fail cleanly
when passing a bad filename to modify. Still need to trap the
non-router environment earlier and restrict its run to the modules(s)
where it is allowed (Currently only "modify")